pyproject.toml
[project]
name = "web-history-search-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "faiss-cpu>=1.10.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.0.2",
    "python-dotenv>=1.1.0",
]
